# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vatel AI REST API",
    "description": "Programmatic access for agents, telephony, and session tokens. Send `Authorization: Bearer <organization API key>` on all requests unless noted. Most error responses return JSON with an `error` string message.",
    "version": "0.2.2"
  },
  "servers": [
    {
      "url": "https://api.vatel.ai",
      "description": "Production"
    }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/v1/session-token": {
      "post": {
        "tags": ["session"],
        "summary": "Generate session token",
        "description": "Short-lived session JWT in `token` to connect to an agent. Send `agent_id` and optional `transport`: `websocket` (default) or `webrtc` (adds `webrtc_token`, `room`, `identity`, `url`). Optional: `version_id`, `prompt`, `first_message`, and `chat` to include the `chat` scope alongside `voice` in the JWT.",
        "operationId": "generateSessionToken",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/GenerateSessionTokenRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session credentials (`token`; with `webrtc`, also `webrtc_token`, `room`, `identity`, `url`).",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionTokenResponse" }
              }
            }
          },
          "400": { "description": "Invalid JSON, disallowed extra properties, missing or invalid `agent_id`, invalid `version_id`, or invalid `transport`." },
          "401": { "description": "Unauthorized" },
          "403": {
            "description": "Agent belongs to another organization",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "404": {
            "description": "Agent not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          },
          "503": {
            "description": "`webrtc` was requested but WebRTC media credentials or endpoint are not configured on the server.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key",
        "description": "Send your organization API key as `Authorization: Bearer <key>`."
      }
    },
    "schemas": {
...
      "GenerateSessionTokenRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": ["agent_id"],
        "properties": {
          "agent_id": {
            "type": "string",
            "format": "uuid",
            "description": "UUID of the agent that will run this voice session."
          },
          "version_id": {
            "type": "string",
            "format": "uuid",
            "description": "Optional published version UUID. When omitted, the server uses the agent's resolved active version."
          },
          "transport": {
            "type": "string",
            "enum": ["websocket", "webrtc"],
            "description": "`websocket`: platform WebSocket session (default when omitted). `webrtc`: WebRTC media room; response adds `room`, `identity`, and `url`."
          },
          "first_message": {
            "type": "string",
            "description": "Optional opening line override for this session (WebSocket session token and WebRTC JWT participant attributes when applicable)."
          },
          "prompt": {
            "type": "string",
            "description": "Optional system prompt override for this session (WebSocket session parameters and WebRTC JWT participant attributes when applicable)."
          },
          "chat": {
            "type": "boolean",
            "description": "When true, the `token` includes the `chat` scope in addition to `voice`, authorizing the chat API."
          }
        }
      },
      "SessionTokenResponse": {
        "type": "object",
        "description": "Session JWT in `token`; for `webrtc`, also `webrtc_token`, `room`, `identity`, and `url`.",
        "required": ["token"],
        "properties": {
          "token": {
            "type": "string",
            "description": "Platform session JWT. `scopes` claim includes `voice`, and `chat` when chat was requested. Use for WebSocket sessions and as `Authorization: Bearer` for the chat endpoint."
          },
          "room": {
            "type": "string",
            "description": "Media room name assigned by the server. Present only when `transport` was `webrtc`."
          },
          "identity": {
            "type": "string",
            "description": "Participant identity assigned by the server for the WebRTC session. Present only when `transport` was `webrtc`."
          },
          "url": {
            "type": "string",
            "description": "WebSocket URL of the media server. Present only when `transport` was `webrtc`."
          },
          "webrtc_token": {
            "type": "string",
            "description": "LiveKit media-room access JWT (microphone subscribe and publish; participant attributes include `agent_id` and, when provided, `version_id`, `prompt`, and `first_message`). Present only when `transport` was `webrtc`."
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Standard error body for most non-2xx JSON responses on this API.",
        "required": ["error"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable message; may include validation detail." }
        }
      }
    }
  }
}